+++ CHANGELOG.md
@@ -20,6 +20,10 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 * 📝 Removed the unsupported release for devices running Android 4 from Google Play, it is no longer
   working properly.
 
+### Next release
+
+* 🔨 On Android 16, when the widget updates, display the correct poster for items.
+
 ### 2025.2.10 - 2025-09-04
 
 * 🔨 Scrolling upcoming and recent tabs should now actually no longer result in a crash in rare